# Speed up __eq__ by generating code