--- a/changelog.d/306.change.rst
+++ b/changelog.d/306.change.rst
@@ -1,0 +1,1 @@
+Equality tests are *much* faster now.