@Tinche 2017-12-03
The speed up looks excellent. So, an evil thought: if the class in question is a dict class, does `self.__dict __ == other.__dict__` fulfills all our requirements? 😈