@hynek 2017-12-03
Where are you getting your ideas? ಠ_ಠ 

(For posterity: no it doesn’t :) — unless class is frozen
And all attributes are part of eq and probably a few more invariants.)