@Tinche 2017-12-03
I admit the idea is probably not usable in practice but I'd like to get to why exactly? All attributes are in `instance.__dict__`. And what does the frozenness of the class have to do with it?

Also consider this:
```
@attr.s
class C:
    a = attr.ib()

>>> a =  C(1)
>>> b = C(1)
>>> a == b
True
>>> a.b = 2
>>> a == b
True
```