@Tinche 2017-12-03
Just to clarify, I'm considering the eligibility of the following `__eq__` implementation (but just for dict classes, obviously):
```
def __eq__(self, other):
    return self.__dict__ == other.__dict__
```