@hynek 2017-12-03
The main reason is that it changes the whole contract between attrs and its users.   Currently we promise we only care about attributes that are explicitly defined.   This approach would compare everything.  Frozenness technically prevents  people from tacking more attributes on instances that we shouldn’t know about.