@hynek 2017-12-04
Now that I’m not on a phone anymore: what I mean is something like this:

```pycon
>>> import attr

>>> @attr.s(cmp=False)
... class C:
...     x = attr.ib()
...     def __eq__(self, other):
...         return self.__dict__ == other.__dict__

>>> C(1)
C(x=1)

>>> c1 = C(1)

>>> c2 = C(1)

>>> c1 == c2
True

>>> c2._cache = {}

>>> c1 == c2
False
```

Leaving aside better ways to achieve that, I’m sure people have done that and breaking backward compat for no real gain doesn’t seem a great way forward to me…