@Tinche 2017-12-04
Yeah, yeah, point taken, I recant my suggestion. Damn users, making it hard for us :(

At this point I'd really like to know if there's any speed difference between a chain of ands and comparing tuples; think this is easy to hack together on this branch?