@Tinche 2017-12-04
Ah I see you did the ands benchmark in another thread, and it's slower? Hm. I guess tuples are the way to go then.