@hynek 2017-12-04
Yeah it’s def slower if the classes are equal.  The difference is in any case marginal, so we can choose what we do.  For now I’d prefer to not break backward compatibility. :)