@hynek 2017-12-04
I’ve stopped using `"".format()` because it’s slow and ugly.  f-strings for lyfe! :)